Testing and refactoring (#11)
## Moving \`VisitorMut\` to \`Traverse\`

Oxide has (at least) two ways implement AST traversal: \`Visitor\`/\`VisitorMut\` and \`Traverse\`.  

The original implementation in the first PR was using \`VisitorMut\` I ran into issues where I needed to refer back to a node's ancestry, which is not easily done using the visitors but is baked right into \`Traverse\`.

## Testing

Testing should be much easier to implement now as all you can to do is create a \`.tsx\` input  file that in \`src/test_input\` and create a test with the matching name and then call the \`assert_valid_transform!\` or \`assert_valid_transform_debug!\` macro in the test's body.  This run will run \`transform\` over the input file, generating Insta snapshots that can be reviewed and accepted for subsequent tests.  

Each exported component will be stored in its own \`.snap\` file that has meaningful name that includes the test name and fully-qualified component id, including the hash.  This makes it very easy to correlate test results with transform output for debugging purposes.  It also allows a single test case to partially succeed in that you can review an accept component snapshots that we created correctly whilst rejecting those that do not.

### Currently Failing Tests

- \`transform::tests::test_example_7\`:  \`App_component_ckEPmXZlub0\` contains extraneous imports that need to be culled.
- \`transform::tests::test_example_9\`:  \`Header_WjUaUQN7Oxg\` contains code that the original snapshot shows as being removed.

## Naming "Improvements"

I have renamed most of the naming that was using \`qwik\` in the name.

## Code Hygiene

\`cargo fix\`'ed all warnings.  Deleted dead code.